QOL changes.

+++ src/CacheData.java
@@ -1,5 +1,5 @@
 public record CacheData(
         String difficulty,
-        int score
+        long score
 ) {
 }
